test_project/test_requirements.txt
Django==1.7.1
Fabric==1.8.2
Pillow==2.6.1
PyJWT==0.3.0
amqp==1.4.6
anyjson==0.3.3
billiard==3.3.0.19
celery==3.1.16
django-braces==1.4.0
django-model-utils==2.2
django-oauth-toolkit==0.7.2
djangorestframework==2.4.4
ecdsa==0.10
factory-boy==2.4.1
futures==2.2.0
kombu==3.0.24
mock==1.0.1
oauthlib==0.7.2
paramiko==1.12.2
pycrypto==2.6.1
python-openid==2.2.5
python-social-auth==0.2.1
python-swiftclient==2.3.1
pytz==2014.9
requests==2.2.1
requests-oauthlib==0.4.2
simplejson==3.3.3
six==1.8.0
stevedore==0.15
twython==3.2.0
virtualenv==1.11.6
virtualenv-clone==0.2.5
virtualenvwrapper==4.3
wsgiref==0.1.2
yak-server==0.1
